+++ src/app-auth/decorator/dashboard-access.decorator.ts
@@ -1,0 +1,4 @@
+import { SetMetadata } from '@nestjs/common';
+
+export const RequireDashboardAccess = (access: string) =>
+  SetMetadata('required_access', access);